feat(tric) skip fixuid script on non-linux OSes
This fix puts a `FIXUID=0` env var that will be parsed from `tric`
custom containers and the `composer` and `codeception` containers used
by the stack to skip the `fixuid` script when not running on Linux.

Default behavior, to make sure CI keeps working is to **not skip**; on
systems that would not require it, the fix simply wasted time, but it's
not harmful.

The fix also adds the `update` command to rebuild images and pull the
latest version of the external ones.

fix #19