fix(tric) handle file modes correctly
fixes #19

This fix is an iteration on the previous one for #19 to make sure file
mode issues are handled correctly in OSes, like Windows or macOS, where
user ID and group ID remapping is performed at the docker daemon level.

The gotcha here is that `docker-compose` will ignore a `user` parameter
set, for a service` to  `":"` to use, instead, the default user.

Previously the stack `user` parameter would be set to `"0:0"` that would
force the use of the `root` user.

If `docker-compose` finds `":"` it will not pass the `--user` flag to
the `docker` binary at all, using the default image user and letting the
user remapping, where supported (macOS and Windows) kick-in.

Furthermore this fixes a number of issues I've found while working on
Linux.