Move commands to separate files
The goal of this PR is to just chunk up the work being done within the switch statement into managable pieces without changing too much. In the future, it may be worth considering switching to Symfony Console.

But that's a lift for another time. The simplicity of this is convenient, though quite old-school procedural.